Proto Snake Game
It follows the grid and works almost well. It does die still when the
snake locks itself. Before fixing this problem, I will make the snake be
able to go free from the grid and in the 3d space. Not extremely
dificult, just need to remmember to do it after finishing the other
tools.